requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
MouseInfo==0.1.3
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
tifffile==2025.2.18
